clear pages
