+++ .gitignore
@@ -996,3 +996,4 @@ $RECYCLE.BIN/
 
 # End of https://www.toptal.com/developers/gitignore/api/macos,linux,windows,vim,visualstudiocode,webstorm,rider,csharp,dotnetcore,aspnetcore,vue,vuejs,visualstudio
 
+chores-client
